pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nicewebrl"
version = "0.0.3"
authors = [
    {name = "Wilka Carvalho", email = "wcarvalho92@gmail.com"},
]
description = "Python library for quickly making interactive RL Apps with NiceGUI"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "nicegui",
    "tortoise-orm",
    "Pillow",
    "jax>=0.4.16",
    "jaxlib>=0.4.16",
    "distrax>=0.1.5",
    "flax>=0.8.2",
    "numpy>=1.26.4,<2",
    "polars",
]

[project.optional-dependencies]
craftax = [
    "craftax",
    "python-dotenv",
    "jax>=0.4.16,<=0.4.22",  # needed for backward compatibility
    "jaxlib>=0.4.16,<=0.4.22",  # needed for backward compatibility
]
jaxmarl = [
    "jaxmarl @ git+https://github.com/FLAIROx/JaxMARL.git",
    "python-dotenv",
    "jax>=0.4.23,<=0.4.38",
    "jaxlib>=0.4.23,<=0.4.38",
]
xland = [
    "xminigrid",
    "tqdm",
]
xland-assistant = [
    "nicegui>=1.4.0",
    "fastapi>=0.68.0",
    "xminigrid>=0.1.0",
    "tortoise-orm>=0.19.0",
    "jax>=0.4.16",
    "jaxlib>=0.4.16",
    "httpx>=0.24.0",
    "uvicorn>=0.15.0",
    "tqdm",
    "google-cloud-storage",
]

[project.urls]
"Homepage" = "https://github.com/wcarvalho/nicewebrl"
"Bug Tracker" = "https://github.com/wcarvalho/nicewebrl/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = []

[tool.setuptools.package-data]
"*" = ["*.js"]